test(KEN-1241): audit pi-session-manager test surfaces
